Add Acc Invitation and FE Pembayaran Borongan

+++ app/Admin/Controllers/GroupOrderPaymentController.php
@@ -8,6 +8,7 @@ use Encore\Admin\Grid;
 use Encore\Admin\Show;
 use App\Models\GroupOrder;
 use App\Models\GroupOrderPayment;
+use Illuminate\Support\Facades\File;
 use Encore\Admin\Controllers\AdminController;
 use App\Admin\AdminExtensions\GroupOrderPaymentExporter;
 
@@ -102,26 +103,28 @@ class GroupOrderPaymentController extends AdminController
         $form->submitted(function (Form $form) {
             $form->ignore('paid_file');
 
-            if(empty($form->model()->get()->toArray())){
-                $filename = md5(request()->file('paid_file')->getClientOriginalName() . time()) . '.' . request()->file('paid_file')->getClientOriginalExtension();
-                request()->file('paid_file')->move(public_path('uploads/payments'), $filename);
+            if(!is_null($form->paid_file)){
+                if(empty($form->model()->get()->toArray())){
+                    $filename = md5(request()->file('paid_file')->getClientOriginalName() . time()) . '.' . request()->file('paid_file')->getClientOriginalExtension();
+                    request()->file('paid_file')->move(public_path('uploads/payments'), $filename);
 
-                $form->model()->paid_file = 'payments'.'/' . $filename;
-            
-                // $form->paid_file = 'payments'.'/' . $filename;
-            }else{
-                $old_file = $form->model()->paid_file;
+                    $form->model()->paid_file = 'payments'.'/' . $filename;
                 
-                if (File::exists(public_path('uploads').$old_file)) {
-                    File::delete(public_path('uploads').$old_file);
+                    // $form->paid_file = 'payments'.'/' . $filename;
+                }else{
+                    $old_file = $form->model()->paid_file;
+                    
+                    if (File::exists(public_path('uploads').$old_file)) {
+                        File::delete(public_path('uploads').$old_file);
+                    }
+
+                    $filename = md5(request()->file('paid_file')->getClientOriginalName() . time()) . '.' . request()->file('paid_file')->getClientOriginalExtension();
+                    request()->file('paid_file')->move(public_path('uploads/payments'), $filename);
+
+                    $form->model()->paid_file = 'payments'.'/' . $filename;
+                    
+                    $form->model()->save();
                 }
-
-                $filename = md5(request()->file('paid_file')->getClientOriginalName() . time()) . '.' . request()->file('paid_file')->getClientOriginalExtension();
-                request()->file('paid_file')->move(public_path('uploads/payments'), $filename);
-
-                $form->model()->paid_file = 'payments'.'/' . $filename;
-                
-                $form->model()->save();
             }
         });
